refactor(frontend): extract helpers from PageManager, PromptBlocksTab, VariantsTab, pipe-demo (#244)
Batch 10: Extract constants, types, and form state into co-located helper files.
- PageManager.tsx: extract SECTION_LABELS, SECTION_ICONS, PAGE_ICONS, types → page-manager-helpers.ts
- PromptBlocksTab.tsx: extract CATEGORY/SCOPE options, badge classes, BlockFormState → prompt-blocks-helpers.ts
- VariantsTab.tsx: extract MOOD options, badge classes, VariantFormState → variants-helpers.ts
- pipe-demo.tsx: extract ChatStep type and CONVERSATION array → pipe-demo-helpers.ts
