+++ frontend/src/app/(platform)/admin/settings/onboarding/_components/PromptBlocksTab.tsx
@@ -31,6 +31,14 @@ import type {
   PromptBlockCategory,
   PromptBlockScope,
 } from '@/types/admin';
+import {
+  CATEGORY_OPTIONS,
+  SCOPE_OPTIONS,
+  CATEGORY_BADGE_CLASSES,
+  SCOPE_BADGE_CLASSES,
+  EMPTY_BLOCK_FORM,
+  type BlockFormState,
+} from './prompt-blocks-helpers';
 import { Badge } from '@/components/ui/badge';
 import { Switch } from '@/components/ui/switch';
 import {
@@ -74,62 +82,6 @@ import {
   BTN_SECONDARY,
   ACTION_TRIGGER,
 } from '../_helpers';
-
-// ─── Constants ───────────────────────────────────────────────────────
-
-const CATEGORY_OPTIONS: { value: PromptBlockCategory; label: string }[] = [
-  { value: 'system', label: 'System' },
-  { value: 'business', label: 'Business' },
-  { value: 'visual', label: 'Visual' },
-  { value: 'section', label: 'Section' },
-  { value: 'rules', label: 'Rules' },
-];
-
-const SCOPE_OPTIONS: { value: PromptBlockScope; label: string }[] = [
-  { value: 'global', label: 'Global' },
-  { value: 'template', label: 'Template' },
-  { value: 'industry', label: 'Industry' },
-];
-
-const CATEGORY_BADGE_CLASSES: Record<PromptBlockCategory, string> = {
-  system: 'border-slate-200 bg-slate-50 text-slate-700',
-  business: 'border-blue-200 bg-blue-50 text-blue-700',
-  visual: 'border-violet-200 bg-violet-50 text-violet-700',
-  section: 'border-emerald-200 bg-emerald-50 text-emerald-700',
-  rules: 'border-amber-200 bg-amber-50 text-amber-700',
-};
-
-const SCOPE_BADGE_CLASSES: Record<PromptBlockScope, string> = {
-  global: 'border-slate-200 bg-slate-50 text-slate-700',
-  template: 'border-teal-200 bg-teal-50 text-teal-700',
-  industry: 'border-indigo-200 bg-indigo-50 text-indigo-700',
-};
-
-// ─── Form state ──────────────────────────────────────────────────────
-
-interface BlockFormState {
-  key: string;
-  label: string;
-  content: string;
-  category: PromptBlockCategory;
-  scope: PromptBlockScope;
-  template: string;
-  industry: string;
-  sort_order: number;
-  is_active: boolean;
-}
-
-const EMPTY_BLOCK_FORM: BlockFormState = {
-  key: '',
-  label: '',
-  content: '',
-  category: 'system',
-  scope: 'global',
-  template: '',
-  industry: '',
-  sort_order: 0,
-  is_active: true,
-};
 
 // ─── Component ───────────────────────────────────────────────────────
